spartan3e-vhdl-psc-setup-master/host/requirements.txt
pyserial
orgasm-package 
sounddevice
numpy